admin_features

+++ app/Http/Controllers/UrlController.php
@@ -6,6 +6,8 @@ use App\Http\Requests\UrlsStoreRequest;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use App\Models\Url;
+use App\Models\User;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Cache;
 use Illuminate\Support\Str;
 use Illuminate\View\View;
@@ -14,7 +16,12 @@ class UrlController extends Controller
 {
     public function index(Request $request):View
     {
-        $urls = Url::where('user_id',$request->user()->id)->get();
+        if($request->user()->is_admin){
+            $urls = Url::all();
+        }
+        else {
+            $urls = Url::where('user_id', $request->user()->id)->get();
+        }
         return view('urls',['urls'=>$urls]);
     }
     public function store(UrlsStoreRequest $request) :RedirectResponse
@@ -29,9 +36,11 @@ class UrlController extends Controller
         return redirect()->back();
 
     }
-    public function destroy(Url $url):RedirectResponse
+    public function destroy(Request $request, Url $url ):RedirectResponse
     {
-        $url->delete();
+        if ($request->user()->is_admin || $request->user()->id === $url->user_id){
+            $url->delete();
+        }
         return redirect()->back();
     }
     public function redirect_counter(string $code)
@@ -41,7 +50,6 @@ class UrlController extends Controller
          return   redirect()->route('urls.index')->with('failure','Такой сслыки не существует');
         }
         $url->update(['count' => $url->count +1]);
-//        dd($url->link);
         return redirect($url->link);
     }
 }
